% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folder.R
\name{create_folder}
\alias{create_folder}
\title{Takes in a string and creates a new folder with that name.}
\usage{
create_folder(folder)
}
\arguments{
\item{folder}{Path (string) of the new directory to be created.}
}
\value{
The input parameter "folder" that indicates the path of the new folder.
}
\description{
String path can contain multiple folders that do not exist yet, separated by "/".
It returns the new path name, in case it needs to be assigned to a variable.
}
\examples{
create_folder("Sub1/Sub2")
}
