% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nth_part.R
\name{get_nth_part}
\alias{get_nth_part}
\title{Extracts the first (or nth) entry from each split.}
\usage{
get_nth_part(x, delimiter, entry_no)
}
\arguments{
\item{x}{A character vector.}

\item{delimiter}{A character to split by.}

\item{entry_no}{A number indicating which part of split to return.}
}
\value{
A character vector containing parts
}
\description{
Extracts the first (or nth) entry from each split.
}
\examples{
x = "2wk.1"   "2wk.2"   "2wk.3"   "2wk.4"   "2wk.5"   "2wk.6"   "2wk.7"   "2wk.8"   "2wk.9"   "2wk.10"
         delimiter = "\\\\."
         entry_no = 1L
         The return value is "2wk"  "2wk"  "2wk"  "2wk"  "2wk"  "2wk"  "2wk"  "2wk"  "2wk"  "2wk"  "2wk"  "2wk"
}
