% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_packages.R
\name{load_packages}
\alias{load_packages}
\title{Install/load packages from library into environment.}
\usage{
load_packages(pkgs)
}
\arguments{
\item{pkgs}{A character vector of package names.}
}
\description{
Checks whether R packages are installed from CRAN and loads
all. Does not yet support installation of BioConductor packages.
}
\examples{
load_packages(c("openxlsx", "dplyr"))
}
