% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xl_to_list.R
\name{read_xl_to_list}
\alias{read_xl_to_list}
\title{Read in Excel file table to list}
\usage{
read_xl_to_list(file, sheet, ...)
}
\arguments{
\item{file}{Name of Excel file (.xlsx), string.}

\item{sheet}{Sheet (name or index) to read in as list.}

\item{...}{Additional parameters passed to openxlsx::read.xlsx}
}
\description{
Each column becomes a vector and a list element, column name becomes list element name.
}
