% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_one.by.one.R
\name{split_one.by.one}
\alias{split_one.by.one}
\title{Split string and iteratively return parts pasted to first part}
\usage{
split_one.by.one(x, delimiter, un_list = T)
}
\arguments{
\item{x}{A character vector.}

\item{delimiter}{A character to split by.}

\item{un_list}{A logical value whether to unlist}
}
\value{

}
\description{
Split string and iteratively return parts pasted to first part
}
\examples{
split_one.by.one("Heatmaps/BY.CORE/TMA.stromal.subtype_all excl HRD excl neo dots Moffitt/Surv", "/")
returns:
}
