% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_df_to_Excel.R
\name{write_df_to_Excel}
\alias{write_df_to_Excel}
\title{Writes a data frame to a worksheet in an Excel workbook.}
\usage{
write_df_to_Excel(sheet_name, my_df, wb, incl_rownames = F)
}
\arguments{
\item{sheet_name}{A string indicating the name of the worksheet.}

\item{my_df}{A data frame.}

\item{wb}{An openxlsx workbook object containing all current worksheets.}

\item{incl_rownames}{Logical indicating whether rownames of the input data frame should be printed, default F.}
}
\value{
The workbook with the new sheet added. Does not have to be reassigned to wb in main.
}
\description{
Writes a data frame to a worksheet in an Excel workbook.
}
\examples{
wb <- openxlsx::createWorkbook() # Create a workbook
my_df <- DNase
write_df_to_Excel("Sheet1", my_df, wb)
openxlsx::saveWorkbook(wb, file="filename.xlsx") # Save wb to file
}
