% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_list_to_Excel.R
\name{write_list_to_Excel}
\alias{write_list_to_Excel}
\title{Writes a list to a worksheet in an Excel workbook.}
\usage{
write_list_to_Excel(sheet_name, my_list, wb)
}
\arguments{
\item{sheet_name}{A string indicating the name of the worksheet.}

\item{my_list}{A named list with vectors (can have varying sizes).}

\item{wb}{An openxlsx workbook object containing all current worksheets.}
}
\value{
The workbook with the new sheet added. Does not have to be reassigned to wb in main.
}
\description{
Writes a list to a worksheet in an Excel workbook.
}
\examples{
wb <- openxlsx::createWorkbook() # Create a workbook
my_list <- list(a=1:3, b=5:12)
write_list_to_Excel("Sheet2", my_list, wb)
openxlsx::saveWorkbook(wb, file="filename.xlsx") # Save wb to file
}
